fw-5518, cleanup search code
- remove unused code
- organize utils now that there are fewer
