+++ firstvoices/backend/search/documents/language_document.py
@@ -1,7 +1,7 @@
 from elasticsearch_dsl import Keyword, Text
 
+from backend.search.constants import ELASTICSEARCH_LANGUAGE_INDEX
 from backend.search.documents.base_document import BaseDocument
-from backend.search.utils.constants import ELASTICSEARCH_LANGUAGE_INDEX
 
 
 class LanguageDocument(BaseDocument):
