+++ firstvoices/backend/search/documents/song_document.py
@@ -1,7 +1,7 @@
 from elasticsearch_dsl import Boolean, Keyword, Text
 
+from backend.search.constants import ELASTICSEARCH_SONG_INDEX
 from backend.search.documents.base_document import BaseSiteEntryWithMediaDocument
-from backend.search.utils.constants import ELASTICSEARCH_SONG_INDEX
 
 
 class SongDocument(BaseSiteEntryWithMediaDocument):
