+++ firstvoices/backend/search/utils/search_term_query.py
@@ -1,6 +1,6 @@
 from elasticsearch_dsl import Q
 
-from backend.search.utils.constants import FUZZY_SEARCH_CUTOFF
+from backend.search.constants import FUZZY_SEARCH_CUTOFF
 
 BASE_BOOST = 1.0  # default value of boost
 EXACT_MATCH_PRIMARY_BOOST = 5
