+++ firstvoices/backend/tests/test_apis/test_search_apis/test_search_querying/test_query_params.py
@@ -2,7 +2,7 @@ import pytest
 from django.contrib.auth.models import AnonymousUser
 
 from backend.models.constants import AppRole, Role, Visibility
-from backend.search.utils.constants import FUZZY_SEARCH_CUTOFF
+from backend.search.constants import FUZZY_SEARCH_CUTOFF
 from backend.search.utils.search_term_query import (
     EXACT_MATCH_OTHER_BOOST,
     EXACT_MATCH_PRIMARY_BOOST,
