+++ firstvoices/backend/tests/test_search_indexing/test_language_indexing.py
@@ -3,12 +3,12 @@ from unittest.mock import patch
 import pytest
 
 from backend.models.constants import Visibility
+from backend.search.constants import ELASTICSEARCH_LANGUAGE_INDEX
 from backend.search.indexing.language_index import (
     LanguageDocumentManager,
     LanguageIndexManager,
     SiteDocumentManager,
 )
-from backend.search.utils.constants import ELASTICSEARCH_LANGUAGE_INDEX
 from backend.tests import factories
 from backend.tests.test_search_indexing.base_indexing_tests import (
     BaseDocumentManagerTest,
