+++ firstvoices/backend/tests/test_search_indexing/test_song_indexing.py
@@ -1,8 +1,8 @@
 import pytest
 
 from backend.models.constants import Visibility
+from backend.search.constants import ELASTICSEARCH_SONG_INDEX
 from backend.search.indexing.song_index import SongDocumentManager, SongIndexManager
-from backend.search.utils.constants import ELASTICSEARCH_SONG_INDEX
 from backend.tests import factories
 from backend.tests.test_search_indexing.base_indexing_tests import (
     BaseDocumentManagerTest,
