+++ firstvoices/backend/views/search_site_entries_views.py
@@ -6,10 +6,7 @@ from drf_spectacular.utils import (
 )
 
 from backend.models import Category, ImportJob
-from backend.search.utils.validators import (
-    get_valid_instance_id,
-    get_valid_starts_with_char,
-)
+from backend.search.validators import get_valid_instance_id, get_valid_starts_with_char
 from backend.views.api_doc_variables import site_slug_parameter
 from backend.views.base_search_entries_views import BaseSearchEntriesViewSet
 from backend.views.base_views import SiteContentViewSetMixin
